debugged client counter
Error is because Verus doesn't know the `client` is the same as the `virtual_alarm.client` and that they are all the same. In general, I don't even want this to be true.

The verification error is correct. It doesn't make sense to count the number of calls in the callback like I'm doing now. I will refactor it so that the Mux is doing the counting since it calls the callback.

```rs
error: precondition not satisfied
    --> capsules/core/src/virtualizers/new_virtual_alarm.rs:1696:27
     |
1618 |             self.client_counter_wf(old(state)),
     |             ---------------------------------- failed precondition
...
1696 |         let fired_count = client.count(Tracked(&mut client_perm));
```